+++ lib/src/deps_used/_exports.dart
@@ -1,0 +1,3 @@
+export 'package:dart_dependency_checker/src/deps_used/deps_used_checker.dart';
+export 'package:dart_dependency_checker/src/deps_used/deps_used_params.dart';
+export 'package:dart_dependency_checker/src/deps_used/deps_used_results.dart';