+++ lib/src/deps_used/deps_used_params.dart
@@ -1,0 +1,10 @@
+import 'package:dart_dependency_checker/src/_shared/params.dart';
+
+/// Params for the [DepsUsedChecker].
+class DepsUsedParams extends BaseParams {
+  const DepsUsedParams({
+    required super.path,
+    super.mainIgnores,
+    super.devIgnores,
+  });
+}